[Player] auto play next episode in set
