+++ views/index.ejs
@@ -11,12 +11,14 @@
   <div class="browser">
     <h1 class="title">VIDEO CAPTAIN</h1>
     <% videos.forEach(([ show, episodes ]) => { %>
-      <h2><%= show.replace('/', '').replace(/\//g, ': ').replace(/^\d\d /, '') %></h2>
-      <section class="show">
-        <% episodes.forEach(({ file }) => { %>
-          <a class="video-thumbnail" href="/watch/<%- encodeURIComponent(file) %>"><%= file.replace(/\.\w+$/, '').replace(/^\d\d /, '') %></a>
-        <% }) %>
-      </section>
+      <details class="show" open>
+        <summary><h2><%= show.replace('/', '').replace(/\//g, ': ').replace(/^\d\d /, '') %></h2></summary>
+        <section class="episodes">
+          <% episodes.forEach(({ file }) => { %>
+            <a class="video-thumbnail" href="/watch/<%- encodeURIComponent(file) %>"><%= file.replace(/\.\w+$/, '').replace(/^\d\d /, '') %></a>
+          <% }) %>
+        </section>
+      </details>
     <% }) %>
   </div>
 </body>
